Default to the first available OPL2/3 hwdep
If a NULL or empty device name is passed in, instead of trying to open
an invalid device, enumerate the available hwdep devices and pick a
plausible looking one.

(This will usually be the first available OPL2 or OPL3 hwdep device.
This is good enough for most people, who'll only have one OPL2/3
compatible soundcard in their system. If, like me, you have two, maybe
you'll be unlucky...)

Ideally, there'll be a better API for enumerating these, and we'd do
something more sensible than string searching, but this is an
improvement for now: in most cases the correct device will be chosen
automatically.