Add an "ioport" driver, which performs direct port I/O
This adds an output module which uses direct port I/O. You'll need to
run it as root, and it may break any other driver you have running, so
be careful. By default, it uses ioperm() and outb() to write to the port
(which compiles down to raw 'out' instructions), but it can be made to
use /dev/port instead by changing a #define in the code.

Enable it by using "ioport:<port>" as the device path, where port is the
FM port for the soundcard (for newer PCI cards, this is usually the
card's base + 50h, for ISA cards, this is usually the AdLib 388h). All
values are in hex, and the default is 388h.