Add an opl2lpt driver using libieee1284
This is a basic opl2lpt driver, for use with Serdaco's parallel-port
OPL2 soundcard:
https://www.serdashop.com/OPL2LPT

To use it, set the device path to "opl2lpt:<name>", where <name> is the
name of the parallel port the OPL2LPT is connected to. This is usually
"parport0".

You'll need to make sure no other driver is using it, so you may need to
"rmmod lp" beforehand.

Building this requires libieee1284:
https://github.com/twaugh/libieee1284
http://cyberelk.net/tim/libieee1284/interface/

Note that this _should_ work with the OPL3LPT as well, but will only
have OPL2 functionality.