Check $OPLHW_DEVICE for a default device path
Instead of always defaulting to the first ALSA device, check to see if
there exists an OPLHW_DEVICE environment variable, and use that if
present.

Passing a device path to oplhw_OpenDevice() will take precedence over
this, but if it is NULL or the empty string, the default device will be
used.

Example:
OPLHW_DEVICE="retrowave:/dev/ttyACM0" oplhw_imfplay "" mymusic.imf