opl2lpt: Add a Linux backend which doesn't require libieee1284
Since there are several distros (e.g., OpenSUSE) which don't seem to
provide libieee1284, let's add our own backend.

This will be preferred, in order to be as consistent as possible between
Linux distros, and to reduce dependencies.

The Linux kernel headers will need to be installed to build this, and
libieee1284 can be3 used instead by passing -DOPLHW_USE_IEEE1284=ON to
cmake.