lpt_linux: Look for device in /dev
There are two LPT backends, one using libieee1284, which looks in /dev
for the device (usually parport0), and one which talks to Linux
directly, and needs the full path to the device file. In order to allow
both to work with the same config string, make the Linux one also look
in the /dev directory if the device is not otherwise found.

Also fix the error detection which used != 0 instead of < 0.