Add basic volume filter support
We need to be able to mute OPL output (and, more importantly, unmute
it), so add a "volume control" feature, which adjusts the OPL stream to
change note volumes.

It isn't perfect yet, but should work.

Basically, we create a "filter device", which is an oplhw_device which
wraps an actual device, modifying note on messages to adjust the volume.

Use:
oplhw_CreateVolumeFilter(device) to create a filter around device.
You can the change the volume with:
oplhw_SetVolume(vol_device, volume) to adjust the volume after-the-fact.
vol_device _must_ be the filter device.

Possible todos:
- Modify the volume (or at least turn off) existing notes.
- Support separate L/R volumes.
- Maybe support SetVolume() on an ALSA device, and adjust the card's
  'Synth' volume channel, if present?