Only export the public API
On Linux systems, by default every symbol is exported in a .so file.
We'd much rather only support a small number as officially exported, so
use -fvisibility=hidden and an OPLHW_API macro to override it for the
public symbols.

This has been tested on Linux, as well as with mingw32, and it seems to
work thus far.

(There's still some work to do to actually _improve_ and _finish_ the
public API, but at least we've got a macro now.)